Move maintenance status summary to overview
